% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_latex.R
\name{export_to_latex}
\alias{export_to_latex}
\title{An Export Function to produce a LaTeX table}
\usage{
export_to_latex(x)
}
\arguments{
\item{x}{Data matrix or Data Frame to be exported into a tex table.}
}
\value{
% latex table generated in R 3.2.4 by xtable 1.8-2 package
% Tue Apr  5 16:40:56 2016
\\begin{table}[ht]
\\centering
\\begin{tabular}{rrll}
\\hline
& c.1..2..3. & c..A....B....C.. & c.T..F..T. \\\\
\\hline
1 &   1 & A & TRUE \\\\
2 &   2 & B & FALSE \\\\
3 &   3 & C & TRUE \\\\
\\hline
\\end{tabular}
\\end{table}
}
\description{
Exports the R data object into a tex file with the same name as in the R working environment (into the current working directory). Automatically detects integers, strings, and numerics automatically formatting accordingly as string or numerical (with significant figures). Allows formatting of spaces in rownames.
}
\examples{
data <- matrix(rnorm(500), 50, 10)
rownames(data) <- 1:50
colnames(data) <- letters[1:10]

export_to_latex(data)

example <- cbind(c(1,2,3), c("A", "B", "C"), c(TRUE, FALSE, TRUE))
export_to_latex(example)
}
\keyword{LaTeX,}
\keyword{dataframe,}
\keyword{matrix,}
\keyword{table}
\keyword{tex,}
