% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_to_latex_loop.R
\name{export_to_latex_loop}
\alias{export_to_latex_loop}
\title{An Export Function to produce LaTeX tables for many objects simultaneously}
\usage{
export_to_latex_loop(object_list)
}
\arguments{
\item{object_list}{string or character vector. Names of objects (data matrices or data frames) to be exported as LaTeX tables with names assigned in the R working environment. This is compatible with the output of ls() and any grep() searches of it.}
}
\description{
Calls \code{\link[R2LaTeX]{export_to_latex}} for every objects given. Exports the R data object into a tex file with the same name as in the R working environment (into the current working directory). Automatically detects integers, strings, and numerics automatically formatting accordingly as string or numerical (with significant figures). Allows formatting of spaces in rownames.
}
\examples{
data <- matrix(rnorm(500), 50, 10)
rownames(data) <- 1:50
colnames(data) <- letters[1:10]

table <- matrix(rnorm(1200), 60, 20)
rownames(data) <- 1:50
colnames(data) <- letters[1:20]

objects <- c("data", "table")

export_to_latex_loop(objects)

}
\keyword{LaTeX,}
\keyword{dataframe,}
\keyword{matrix,}
\keyword{table}
\keyword{tex,}
